Maintenance #CHANGE: Split up DualityEditorApp.cs into one file per class.
