+++ DualityEditor/DualityEditor.csproj
@@ -197,6 +197,9 @@
     </Compile>
     <Compile Include="Properties\GeneralResNames.cs" />
     <Compile Include="Utility\BuildHelper.cs" />
+    <Compile Include="Utility\HighlightMode.cs" />
+    <Compile Include="Utility\AutosaveFrequency.cs" />
+    <Compile Include="Utility\SelectMode.cs" />
     <Compile Include="Utility\Extensions\ExtMethodsDualityFont.cs" />
     <Compile Include="Utility\Extensions\ExtMethodsPixelData.cs" />
     <Compile Include="Utility\Extensions\ExtMethodsBitmap.cs" />
