+++ DualityEditor/DualityEditorApp.cs
@@ -26,38 +26,6 @@ using WeifenLuo.WinFormsUI.Docking;
 
 namespace Duality.Editor
 {
-	public enum SelectMode
-	{
-		Set,
-		Append,
-		Toggle
-	}
-
-	[Flags]
-	public enum HighlightMode
-	{
-		None		= 0x0,
-
-		/// <summary>
-		/// Highlights an objects conceptual representation, e.g. flashing its entry in an object overview.
-		/// </summary>
-		Conceptual	= 0x1,
-		/// <summary>
-		/// Highlights an objects spatial location, e.g. focusing it spatially in a scene view.
-		/// </summary>
-		Spatial		= 0x2,
-
-		All			= Conceptual | Spatial
-	}
-
-	public enum AutosaveFrequency
-	{
-		Disabled,
-		TenMinutes,
-		ThirtyMinutes,
-		OneHour
-	}
-
 	public static class DualityEditorApp
 	{
 		public	const	string	EditorLogfilePath		= "logfile_editor.txt";
