+++ DualityEditor/Utility/AutosaveFrequency.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Duality.Editor
+{
+	public enum AutosaveFrequency
+	{
+		Disabled,
+		TenMinutes,
+		ThirtyMinutes,
+		OneHour
+	}
+}
