+++ DualityEditor/Utility/SelectMode.cs
@@ -1,0 +1,11 @@
+﻿using System;
+
+namespace Duality.Editor
+{
+	public enum SelectMode
+	{
+		Set,
+		Append,
+		Toggle
+	}
+}
